create-to-do-list
A new task should be added to the list when the "+" button is clicked.
Clicking the "Done" button should strike through the task and update its status.
Clicking the "Delete" button should remove the task from the list.
The count of completed tasks should be accurately displayed as "X tasks done out of Y."